Remover componente problemático

+++ src/Lib/MCART.Windows/PInvoke/DwmApi.cs
@@ -55,7 +55,7 @@ internal partial class DwmApi
     internal static partial int DwmGetWindowAttribute(IntPtr hwnd, DwmWindowAttribute dwAttribute, [MarshalAs(UnmanagedType.Bool)] out bool pvAttribute, int cbAttribute);
 
     [LibraryImport("dwmapi.dll")] 
-    internal static partial int DwmSetWindowAttribute(IntPtr hwnd, DwmWindowAttribute dwAttribute, ref nint pvAttribute, nint cbAttribute);
+    internal static partial int DwmSetWindowAttribute(IntPtr hwnd, DwmWindowAttribute dwAttribute, ref IntPtr pvAttribute, uint cbAttribute);
 
     [LibraryImport("dwmapi.dll")]
     internal static partial int DwmGetColorizationColor(out uint pcrColorization, [MarshalAs(UnmanagedType.Bool)] out bool pfOpaqueBlend);
@@ -72,7 +72,7 @@ internal partial class DwmApi
     internal static extern int DwmGetWindowAttribute(IntPtr hwnd, DwmWindowAttribute dwAttribute, out bool pvAttribute, int cbAttribute);
 
     [DllImport("dwmapi.dll")]
-    internal static extern int DwmSetWindowAttribute(IntPtr hwnd, DwmWindowAttribute dwAttribute, ref nint pvAttribute, int cbAttribute);
+    internal static extern int DwmSetWindowAttribute(IntPtr hwnd, DwmWindowAttribute dwAttribute, ref uint pvAttribute, int cbAttribute);
 
     [DllImport("dwmapi.dll")]
     internal static extern int DwmGetColorizationColor(out uint pcrColorization, [MarshalAs(UnmanagedType.Bool)] out bool pfOpaqueBlend);
